requirements.txt
autopep8>=1.6.0
click>=8.0.3
coverage>=5.5
Faker>=11.3.0
flask-cors>=3.0.10
flask-FormRequest>=0.0.31
Flask>=2.0
Flask-Migrate>=3.1.0
Flask-Seeder>=1.2.0
Flask-SQLAlchemy>=2.5.1
gunicorn>=20.1
redis>=3.5.3
psycopg2>=2.9.1
pyjwt>=2.3.0
pytest>=6.2
python-dotenv>=0.19.0
pytz>=2022.1
